Updated the code and modified with clean-architecture with MVVM.
