+++ app/src/main/java/com/alpha/myeyecare/MyApplication.kt
@@ -4,6 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class MyApplication: Application() {
-
-}
+class MyApplication: Application()
