+++ app/src/main/java/com/alpha/myeyecare/common/constants/AppDestinations.kt
@@ -1,0 +1,9 @@
+package com.alpha.myeyecare.common.constants
+
+object AppDestinations {
+    const val HOME_SCREEN = "home"
+    const val SPLASH_SCREEN = "splash_screen"
+    const val EYE_CARE_REMINDER_SCREEN = "eye_care"
+    const val WATER_REMINDER_SCREEN = "water_reminder"
+    const val USER_SUGGESTION_SCREEN = "user_suggestions_screen"
+}
