+++ app/src/main/java/com/alpha/myeyecare/common/constants/ReminderTypes.kt
@@ -1,0 +1,7 @@
+package com.alpha.myeyecare.common.constants
+
+object ReminderTypes {
+    const val UNDEFINED = "undefined"
+    const val EYE_REMINDER = "eye_reminder"
+    const val DRINKING_REMINDER = "drinking_reminder"
+}
