+++ app/src/main/java/com/alpha/myeyecare/di/RemoteModule.kt
@@ -1,0 +1,21 @@
+package com.alpha.myeyecare.di
+
+import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object RemoteModule {
+
+    @Provides
+    @Singleton
+    fun provideFirebaseDb(): FirebaseFirestore {
+        return Firebase.firestore
+    }
+}
