+++ app/src/main/java/com/alpha/myeyecare/domain/model/DayOfWeek.kt
@@ -1,0 +1,6 @@
+package com.alpha.myeyecare.domain.model
+
+enum class DayOfWeek(val shortName: String, val fullName: String) {
+    MON("M", "Monday"), TUE("T", "Tuesday"), WED("W", "Wednesday"),
+    THU("T", "Thursday"), FRI("F", "Friday"), SAT("S", "Saturday"), SUN("S", "Sunday")
+}
