+++ app/src/main/java/com/alpha/myeyecare/domain/model/Suggestion.kt
@@ -1,0 +1,16 @@
+package com.alpha.myeyecare.domain.model
+
+import com.google.firebase.firestore.ServerTimestamp
+import java.util.Date
+
+data class Suggestion(
+    val id: String? = null,
+    val name: String = "",
+    val email: String = "",
+    val text: String = "",
+    @ServerTimestamp
+    val timestamp: Date? = null,
+    val status: String = "new"
+) {
+    constructor() : this(null, "", "", "", null, "new")
+}
