+++ app/src/main/java/com/alpha/myeyecare/domain/repository/SuggestionRepository.kt
@@ -1,0 +1,7 @@
+package com.alpha.myeyecare.domain.repository
+
+import com.alpha.myeyecare.domain.model.Suggestion
+
+interface SuggestionRepository {
+    suspend fun addSuggestion(suggestion: Suggestion): Result<String>
+}
