+++ app/src/main/java/com/alpha/myeyecare/domain/useCases/GetReminderDetailsUserCase.kt
@@ -1,0 +1,14 @@
+package com.alpha.myeyecare.domain.useCases
+
+import com.alpha.myeyecare.domain.model.ReminderDetails
+import com.alpha.myeyecare.domain.repository.ReminderRepository
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+class GetReminderDetailsUserCase @Inject constructor(
+    private val repository: ReminderRepository
+) {
+    fun invoke(id: String): Flow<ReminderDetails> {
+        return repository.getReminderDetails(id)
+    }
+}
