--- app/src/main/java/com/alpha/myeyecare/model/AppDestinations.kt
@@ -1,10 +1,0 @@
-package com.alpha.myeyecare.model
-
-
-object AppDestinations {
-    const val HOME_SCREEN = "home"
-    const val SPLASH_SCREEN = "splash_screen"
-    const val EYE_CARE_REMINDER_SCREEN = "eye_care"
-    const val WATER_REMINDER_SCREEN = "water_reminder"
-    const val USER_SUGGESTION_SCREEN = "user_suggestions_screen"
-}
