--- app/src/main/java/com/alpha/myeyecare/model/ReminderTypes.kt
@@ -1,7 +1,0 @@
-package com.alpha.myeyecare.model
-
-
-object ReminderTypes {
-    const val EYE_REMINDER = "eye_reminder"
-    const val DRINKING_REMINDER = "drinking_reminder"
-}
