--- app/src/main/java/com/alpha/myeyecare/model/SuggestionSubmissionUiState.kt
@@ -1,9 +1,0 @@
-package com.alpha.myeyecare.model
-
-// --- UI States for different screens/features ---
-sealed class SuggestionSubmissionUiState {
-    object Idle : SuggestionSubmissionUiState()
-    object Loading : SuggestionSubmissionUiState()
-    data class Success(val message: String) : SuggestionSubmissionUiState()
-    data class Error(val message: String) : SuggestionSubmissionUiState()
-}
