--- app/src/main/java/com/alpha/myeyecare/repository/ISuggestionRepository.kt
@@ -1,7 +1,0 @@
-package com.alpha.myeyecare.repository
-
-import com.alpha.myeyecare.model.Suggestion
-
-interface ISuggestionRepository {
-    suspend fun addSuggestion(suggestion: Suggestion): Result<String>
-}
